gst-plugins-ugly: Autospec creation for update from version 1.16.1 to version 1.16.2
commit 4b2943ee2788331e4000f0f7e226cd5ce2c135a4
Author: Tim-Philipp Müller <tim@centricular.com>
Date:   Tue Dec 3 11:07:52 2019 +0000

    Release 1.16.2

commit 8774d8c5a46d58480cb653a55a6c25d68dd99b62
Author: Tim-Philipp Müller <tim@centricular.com>
Date:   Tue Dec 3 11:07:52 2019 +0000

    Update docs

The latest bug-fix release in the 1.16 series is 1.16.2 and was released
on 3 December 2019.

See https://gstreamer.freedesktop.org/releases/1.16/ for the latest
version of this document.

_Last updated: Tuesday 03 December 2019, 08:00 UTC (log)_

Introduction

The GStreamer team is proud to announce a new major feature release in
the stable 1.x API series of your favourite cross-platform multimedia
framework!

(NEWS truncated at 15 lines)