gst-plugins-ugly: Autospec creation for update from version 1.18.0 to version 1.18.1
commit 720672eed30b3be47b2f26d67554786c0d3693ad
Author: Tim-Philipp Müller <tim@centricular.com>
Date:   Mon Oct 26 11:12:33 2020 +0000

    Release 1.18.1

commit 03122d3d1454adacd52afa5b03951d03d11044c6
Author: Tim-Philipp Müller <tim@centricular.com>
Date:   Tue Sep 8 17:44:20 2020 +0100

    Back to development

The latest bug-fix release in the 1.18 series is 1.18.1 and was released
on 26 October 2020.

See https://gstreamer.freedesktop.org/releases/1.18/ for the latest
version of this document.

Last updated: Monday 26 October 2020, 11:00 UTC (log)

Introduction

The GStreamer team is proud to announce a new major feature release in
the stable 1.x API series of your favourite cross-platform multimedia
framework!

As always, this release is again packed with many new features, bug

(NEWS truncated at 15 lines)