gst-plugins-ugly: Autospec creation for update from version 1.18.1 to version 1.18.2
commit 45c7522506ff49fcde0933c65c1eba48e1df7500
Author: Tim-Philipp Müller <tim@centricular.com>
Date:   Sun Dec 6 13:23:19 2020 +0000

    Release 1.18.2

commit 2feb9715545775462cc6b6df60c9bdc45c4a50d4
Author: Tim-Philipp Müller <tim@centricular.com>
Date:   Tue Oct 27 12:34:22 2020 +0000

    Back to development

The latest bug-fix release in the 1.18 series is 1.18.2 and was released
on 6 December 2020.

See https://gstreamer.freedesktop.org/releases/1.18/ for the latest
version of this document.

Last updated: Monday 26 October 2020, 11:00 UTC (log)

Introduction

The GStreamer team is proud to announce a new major feature release in
the stable 1.x API series of your favourite cross-platform multimedia
framework!

As always, this release is again packed with many new features, bug

(NEWS truncated at 15 lines)