gst-plugins-ugly: Autospec creation for update from version 1.18.2 to version 1.18.3
commit ad60e5463f591caa1c8470f180fd0ef8fce4a802
Author: Tim-Philipp Müller <tim@centricular.com>
Date:   Wed Jan 13 21:09:14 2021 +0000

    Release 1.18.3

commit a9e133420e768f870fb560fa58a60000acae56eb
Author: Tim-Philipp Müller <tim@centricular.com>
Date:   Sun Dec 6 23:56:52 2020 +0000

    Back to development

The latest bug-fix release in the 1.18 series is 1.18.3 and was released
on 13 January 2021.

See https://gstreamer.freedesktop.org/releases/1.18/ for the latest
version of this document.

Last updated: Wednesday 13 January 2021, 20:00 UTC (log)

Introduction

The GStreamer team is proud to announce a new major feature release in
the stable 1.x API series of your favourite cross-platform multimedia
framework!

As always, this release is again packed with many new features, bug

(NEWS truncated at 15 lines)