gst-plugins-ugly: Autospec creation for update from version 1.18.5 to version 1.18.6
commit 8f97a05e1d2cb3cd2548cde2e70ea6d335040f59
Author: Tim-Philipp Müller <tim@centricular.com>
Date:   Wed Feb 2 15:07:08 2022 +0000

    Release 1.18.6

commit 6eabe7821280afc11b7b0d66d85132ff8251adda
Author: Tim-Philipp Müller <tim@centricular.com>
Date:   Thu Sep 9 00:12:42 2021 +0100

    Back to development

The latest bug-fix release in the 1.18 series is 1.18.6 and was released
on 2 February 2022.

See https://gstreamer.freedesktop.org/releases/1.18/ for the latest
version of this document.

Last updated: Wednesday 2 February 2022, 11:30 UTC (log)

Introduction

The GStreamer team is proud to announce a new major feature release in
the stable 1.x API series of your favourite cross-platform multimedia
framework!

As always, this release is again packed with many new features, bug

(NEWS truncated at 15 lines)